requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Django==4.0.5
django-environ==0.8.1
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-http-client==3.3.7
requests==2.28.0
sendgrid==6.9.7
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
